[REFACTOR] : 컴포넌트 분리 (#53)
